getTracerData.py
A simple script to read Epever/Triton solar charge controller realtime data. Run this script to test it works, when its functioning correctly
you redirect the output to a file with
./getTracerData.py > solarData.txt

Note: to be read by Prometheus/Grafana this data needs to be formatted differently. I will upload the code to do that ASAP.